refactor event types and state handling: - Change EventType enum type to uint32_t and update values to align with MainStateIdList. - Update StateInf and all state classes to use uint32_t for eventListener. - Refactor IdleState and LedState event handling to extract event codes from the new event format

+++ RTOS_STM32/apl/menu/MenuStateMachine.cpp
@@ -10,7 +10,7 @@ namespace Menu
     ///polling for event
     if(m_isNewEvent_bo)
     {
-      m_currentState->eventListener(static_cast<uint8_t>(m_rawEventId_u32));
+      m_currentState->eventListener(m_rawEventId_u32);
       m_isNewEvent_bo = false;
     }
